fixed add button
