+++ src/components/Home/Category/CategoryContent/ButtonAdd/ButtonAdd.scss
@@ -25,14 +25,14 @@
     &:before {
       height: 1em;
       width: 0.2em;
-      animation: rotateBefore 6s ease-in forwards;
+      // animation: rotateBefore 6s ease-in forwards;
       // animation: positionYbefore 2s ease-in forwards;
     }
 
     &:after {
       height: 0.2em;
       width: 1em;
-      animation: rotateAfter 6s ease-in forwards;
+      // animation: rotateAfter 6s ease-in forwards;
       // animation: positionXafter 2s ease-in forwards;
       // animation: positionYafter 2s ease-in forwards;
     }
@@ -48,28 +48,7 @@
     //   top: 20%;
     //   left: 50%;
     // }
-    @keyframes rotateBefore {
-      from {
-        transform: rotate(0deg);
-        top: 50%;
-      }
-      to {
-        transform: rotate(-45deg);
-        top: 20%;
-      }
-    }
-    @keyframes rotateAfter {
-      from {
-        transform: rotate(0deg);
-        top: 50%;
-        left: 50%;
-      }
-      to {
-        top: 60%;
-        left: 32.5%;
-        transform: rotate(-45deg);
-      }
-    }
+
     // @keyframes positionYbefore {
     //   from {
     //     top: 50%;
@@ -98,15 +77,72 @@
     // }
   }
 }
-// .plus-button:after:onclick {
-//   transform: rotate(-45deg);
-//   -webkit-transform: rotate(-45deg);
-//   top: 60%;
-//   left: 32.5%;
-// }
-// .plus-button:before:onclick {
-//   transform: rotate(-45deg);
-//   -webkit-transform: rotate(-45deg);
-//   top: 20%;
-//   left: 50%;
-// }
+.rotated {
+  border: 1px solid $mustard;
+  background-color: $mustard;
+  font-size: 20px;
+  height: 2.5em;
+  width: 2.5em;
+  border-radius: 999px;
+  position: relative;
+
+  &:after,
+  &:before {
+    content: "";
+    display: block;
+    background-color: white;
+    position: absolute;
+    // transform: translate(-50%, -50%);
+
+    // transition: transform 1s ease-in;
+  }
+
+  &:before {
+    height: 1em;
+    width: 0.2em;
+    animation: rotateBefore 2s ease-in forwards;
+    // background-color: blue;
+    // animation: positionYbefore 2s ease-in forwards;
+    // transform: rotate(-45deg);
+    // -webkit-transform: rotate(-45deg);
+    // top: 20%;
+    // left: 50%;
+  }
+
+  &:after {
+    height: 0.2em;
+    width: 1em;
+    animation: rotateAfter 2s ease-in forwards;
+    // background-color: red;
+    // transform: rotate(-45deg);
+    // -webkit-transform: rotate(-45deg);
+    // top: 60%;
+    // left: 32.5%;
+  }
+}
+@keyframes rotateBefore {
+  from {
+    transform: rotate(0deg);
+    transform-origin: 50%, 50%;
+    top: 50%;
+    left: 50%;
+  }
+  to {
+    transform: rotate(-45deg);
+    transform-origin: 50%, 50%;
+    top: 20%;
+    left: 50%;
+  }
+}
+@keyframes rotateAfter {
+  from {
+    transform: rotate(0deg);
+    top: 50%;
+    left: 50%;
+  }
+  to {
+    top: 60%;
+    left: 33%;
+    transform: rotate(-45deg);
+  }
+}
