+++ src/components/Home/Category/CategoryContent/ButtonAdd/tests.js
@@ -1,0 +1,17 @@
+// import React, { useState } from "react";
+// import "./ButtonAdd.scss";
+
+// const ButtonAdd = () => {
+//   const [arrowRotated, rotateArrow] = useState(false);
+//   const handleClick = () => {
+//     rotateArrow(!arrowRotated);
+//   };
+//   return (
+//     <div className="btn">
+//       <button className="plus-button" onClick={handleClick}></button>
+//       <input className="add-idea"></input>
+//     </div>
+//   );
+// };
+
+// export default ButtonAdd;
